Move WizData from models.py to coordinator.py
