+++ homeassistant/components/wiz/binary_sensor.py
@@ -18,8 +18,8 @@ from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import WizConfigEntry
 from .const import DOMAIN, SIGNAL_WIZ_PIR
+from .coordinator import WizData
 from .entity import WizEntity
-from .models import WizData
 
 OCCUPANCY_UNIQUE_ID = "{}_occupancy"
