+++ homeassistant/components/wiz/light.py
@@ -23,8 +23,8 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import WizConfigEntry
+from .coordinator import WizData
 from .entity import WizToggleEntity
-from .models import WizData
 
 RGB_WHITE_CHANNELS_COLOR_MODE = {1: ColorMode.RGBW, 2: ColorMode.RGBWW}
