--- homeassistant/components/wiz/models.py
@@ -1,18 +1,0 @@
-"""WiZ integration models."""
-
-from __future__ import annotations
-
-from dataclasses import dataclass
-
-from pywizlight import wizlight
-
-from .coordinator import WizCoordinator
-
-
-@dataclass
-class WizData:
-    """Data for the wiz integration."""
-
-    coordinator: WizCoordinator
-    bulb: wizlight
-    scenes: list
