+++ homeassistant/components/wiz/number.py
@@ -18,8 +18,8 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import WizConfigEntry
+from .coordinator import WizData
 from .entity import WizEntity
-from .models import WizData
 
 
 @dataclass(frozen=True, kw_only=True)
