+++ homeassistant/components/wiz/sensor.py
@@ -17,8 +17,8 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import WizConfigEntry
+from .coordinator import WizData
 from .entity import WizEntity
-from .models import WizData
 
 SENSORS: tuple[SensorEntityDescription, ...] = (
     SensorEntityDescription(
